refactor(hutch): extract shared loginAgent into test-app.ts
Six test files duplicated an identical loginAgent helper with
inconsistent type annotations (TestAppResult vs TestAppHarness,
inline vs top-level Server import). Move the function into
test-app.ts — where both types and the Server import already
live — and re-export it so each test file just imports it.
