+++ projects/hutch/src/runtime/web/pages/export/export.route.test.ts
@@ -1,25 +1,11 @@
-import type { Server } from "node:http";
 import { JSDOM } from "jsdom";
 import request from "supertest";
-import { useTestServer, type TestAppHarness } from "../../../test-app";
+import { useTestServer, loginAgent } from "../../../test-app";
 
 import {
 	TEST_APP_ORIGIN,
 	createDefaultTestAppFixture,
 } from "@packages/test-fixtures";
-
-async function loginAgent(
-	server: Server,
-	auth: TestAppHarness["auth"],
-) {
-	await auth.createUser({ email: "test@example.com", password: "password123" });
-	const agent = request.agent(server);
-	await agent
-		.post("/login")
-		.type("form")
-		.send({ email: "test@example.com", password: "password123" });
-	return agent;
-}
 
 const useApp = useTestServer();
