+++ projects/hutch/src/runtime/web/pages/queue/queue.freshness.route.test.ts
@@ -1,22 +1,11 @@
-import request from "supertest";
 import { initRefreshArticleIfStale } from "@packages/test-fixtures/providers/article-freshness";
 import type { PublishRefreshArticleContent } from "@packages/test-fixtures/providers/events";
 import type { PublishUpdateFetchTimestamp } from "@packages/test-fixtures/providers/events";
-import { useTestServer, type TestAppResult } from "../../../test-app";
+import { useTestServer, loginAgent } from "../../../test-app";
 import {
 	TEST_APP_ORIGIN,
 	createDefaultTestAppFixture,
 } from "@packages/test-fixtures";
-
-async function loginAgent(server: import("node:http").Server, auth: TestAppResult["auth"]) {
-	await auth.createUser({ email: "test@example.com", password: "password123" });
-	const agent = request.agent(server);
-	await agent
-		.post("/login")
-		.type("form")
-		.send({ email: "test@example.com", password: "password123" });
-	return agent;
-}
 
 const useApp = useTestServer();
